clean: made "filter group" and "notification column" their own component
massively clears up notification index
