+++ client/src/components/Profile/UserNotificationFeed/index.tsx
@@ -7,12 +7,8 @@ import {
   markAllNotificationsRead,
   queryNotifications,
 } from "queries/notifications";
-import LoadingBlocks from "components/Artist/LoadingBlocks";
-import usePagination from "utils/usePagination";
-import NotificationFeedItem from "./NotificationFeedItem";
 import React from "react";
-
-const pageSize = 30;
+import NotificationColumn from "./NotificationColumn";
 
 const FOLLOW_TYPES: Notification["notificationType"][] = [
   "NEW_ARTIST_ALBUM",
@@ -27,147 +23,6 @@ const ACTIVITY_TYPES: Notification["notificationType"][] = [
   "LABEL_ADDED_ARTIST",
   "FUNDRAISER_PLEDGE_CHARGED",
 ];
-
-const FilterGroup: React.FC<{
-  legend: string;
-  name: string;
-  options: { value: string; label: string }[];
-  value: string;
-  onChange: (value: string) => void;
-  compact?: boolean;
-}> = ({ legend, name, options, value, onChange, compact }) => (
-  <fieldset className="border-none p-0 m-0 mb-4">
-    <legend className="sr-only">{legend}</legend>
-    <div className="flex flex-wrap gap-1">
-      {options.map((opt) => (
-        <label key={opt.value} className="cursor-pointer">
-          <input
-            type="radio"
-            name={name}
-            value={opt.value}
-            checked={value === opt.value}
-            onChange={() => onChange(opt.value)}
-            className="sr-only peer"
-          />
-          <span
-            className={[
-              "inline-block rounded-full border transition-colors select-none peer-focus-visible:ring-2 peer-focus-visible:ring-(--mi-primary-color) peer-focus-visible:ring-offset-1",
-              compact ? "text-xs px-2 py-0.5" : "text-xs px-2.5 py-1",
-              value === opt.value
-                ? "border-(--mi-primary-color) text-(--mi-primary-color) font-semibold"
-                : "border-(--mi-darken-x-background-color) text-(--mi-light-foreground-color) hover:text-(--mi-normal-foreground-color) hover:border-(--mi-normal-foreground-color)",
-            ].join(" ")}
-          >
-            {opt.label}
-          </span>
-        </label>
-      ))}
-    </div>
-  </fieldset>
-);
-
-const NotificationColumn: React.FC<{
-  title: string;
-  userId: number;
-  baseTypes: Notification["notificationType"][];
-  filterOptions: {
-    value: string;
-    label: string;
-    types: Notification["notificationType"][];
-  }[];
-  filterName: string;
-  compact?: boolean;
-}> = ({ title, userId, baseTypes, filterOptions, filterName, compact }) => {
-  const { t } = useTranslation("translation", { keyPrefix: "notifications" });
-  const [filter, setFilter] = React.useState("all");
-  const { page, resetPage, PaginationComponent } = usePagination({
-    pageSize,
-    pageParam: filterName,
-  });
-
-  const activeTypes =
-    filter === "all"
-      ? baseTypes
-      : (filterOptions.find((o) => o.value === filter)?.types ?? baseTypes);
-
-  const { data, isPending } = useQuery(
-    queryNotifications(userId, {
-      skip: pageSize * page,
-      take: pageSize,
-      notificationType: activeTypes,
-    })
-  );
-
-  const results = data?.results ?? [];
-
-  const unreadIdsAtMount = React.useRef<Set<string> | null>(null);
-  if (unreadIdsAtMount.current === null && data) {
-    unreadIdsAtMount.current = new Set(
-      results.filter((n) => !n.isRead).map((n) => n.id)
-    );
-  }
-
-  React.useEffect(() => {
-    unreadIdsAtMount.current = null;
-  }, [filter]);
-
-  const separatorIndex = unreadIdsAtMount.current
-    ? results.findIndex((n) => !unreadIdsAtMount.current!.has(n.id))
-    : -1;
-
-  const options = [{ value: "all", label: t("filterAll") }, ...filterOptions];
-
-  return (
-    <div className="flex-1 min-w-0">
-      {compact ? (
-        <p className="text-sm font-semibold uppercase tracking-[0.08em] text-(--mi-light-foreground-color) mb-3">
-          {title}
-        </p>
-      ) : (
-        <h2 className="text-lg font-bold text-(--mi-normal-foreground-color) mb-3">
-          {title}
-        </h2>
-      )}
-      <FilterGroup
-        legend={t("filterLegend", { column: title })}
-        name={filterName}
-        options={options}
-        value={filter}
-        onChange={(value) => {
-          setFilter(value);
-          resetPage();
-        }}
-        compact={compact}
-      />
-      {isPending && (
-        <LoadingBlocks height={compact ? "3rem" : "4rem"} rows={5} />
-      )}
-      <ul className="list-none">
-        {results.map((notification, index) => (
-          <React.Fragment key={notification.id}>
-            {separatorIndex > 0 && index === separatorIndex && (
-              <li className="flex items-center gap-3 my-6 text-sm text-(--mi-light-foreground-color)">
-                <div className="flex-1 h-px bg-(--mi-darken-x-background-color)" />
-                <span>{t("earlierNotifications")}</span>
-                <div className="flex-1 h-px bg-(--mi-darken-x-background-color)" />
-              </li>
-            )}
-            <NotificationFeedItem
-              notification={notification}
-              compact={compact}
-            />
-          </React.Fragment>
-        ))}
-        {!isPending && results.length === 0 && (
-          <li className="text-sm text-(--mi-light-foreground-color) py-4">
-            {t("noNotifications")}
-          </li>
-        )}
-      </ul>
-      <PaginationComponent amount={results.length} total={data?.total} />
-    </div>
-  );
-};
 
 const UserNotificationFeed = () => {
   const { t } = useTranslation("translation", { keyPrefix: "notifications" });
